+++ src/main/java/com/y4kstudios/pytestimp/PyTestImpService.kt
@@ -4,10 +4,7 @@ import ca.szc.configparser.Ini
 import ca.szc.configparser.exceptions.NoOptionError
 import ca.szc.configparser.exceptions.NoSectionError
 import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
-import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.ServiceManager
-import com.intellij.openapi.components.State
-import com.intellij.openapi.components.Storage
+import com.intellij.openapi.components.*
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.roots.ProjectRootManager
@@ -15,6 +12,13 @@ import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.newvfs.BulkFileListener
 import com.intellij.openapi.vfs.newvfs.events.VFileEvent
+import com.intellij.psi.PsiFileFactory
+import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.util.execution.ParametersListUtil
+import com.jetbrains.commandInterface.commandLine.CommandLineLanguage
+import com.jetbrains.commandInterface.commandLine.psi.CommandLineArgument
+import com.jetbrains.commandInterface.commandLine.psi.CommandLineFile
+import com.jetbrains.python.sdk.poetry.PY_PROJECT_TOML
 import org.tomlj.Toml
 import org.tomlj.TomlParseResult
 import java.io.BufferedReader
@@ -64,10 +68,10 @@ class PyTestImpService(val project: Project) : PersistentStateComponent<PyTestIm
             refreshPytestConfig(file)
         }
 
-    var pytestConfig: PyTestConfig? = PyTestConfig.parse(null)
+    var pytestConfig: PyTestConfig? = PyTestConfig.parse(null, project)
 
     fun refreshPytestConfig(file: VirtualFile?) {
-        val newPytestConfig = PyTestConfig.parse(file)
+        val newPytestConfig = PyTestConfig.parse(file, project)
 
         if (newPytestConfig != pytestConfig) {
             pytestConfig = newPytestConfig
@@ -209,29 +213,66 @@ fun convertWildcardPatternsStringToRegex(patterns: String, withDashes: Boolean):
     convertWildcardPatternsToRegex(patterns.split(Regex(" +")), withDashes)
 
 
+data class PytestLoadPlugin(
+    val name: String,
+    val excluded: Boolean = false,
+)
+
+
 /**
  * Interface for providing access to a pytest config file (pytest.ini file or pyproject.toml section)
  */
-abstract class PyTestConfig {
+abstract class PyTestConfig(val project: Project) {
     protected open val pythonClassesRaw: String? = null
     protected open val pythonFunctionsRaw: String? = null
 
     open val pythonClasses: Regex by lazy { convertWildcardPatternsStringToRegex(pythonClassesRaw ?: DEFAULT_PYTHON_CLASSES, true) }
     open val pythonFunctions: Regex by lazy { convertWildcardPatternsStringToRegex(pythonFunctionsRaw ?: DEFAULT_PYTHON_FUNCTIONS, false) }
 
+    /** Split/shlex'd extra py.test command-line options, from config's `addopts` */
+    open val cmdlineOpts: List<String> = emptyList()
+
+    /**
+     * String representation of py.test command-line invocation
+     * WARNING: NOT INTENDED FOR EXECUTION — see {@link #com.intellij.execution.configurations.GeneralCommandLine.getCommandLineString()}
+     */
+    open val cmdlineString: String by lazy { ParametersListUtil.join(cmdlineOpts) }
+
+    /** Parsed CommandLine PsiFile */
+    open val cmdlineFile: CommandLineFile by lazy {
+        val psiFileFactory = project.service<PsiFileFactory>()
+        psiFileFactory.createFileFromText(CommandLineLanguage.INSTANCE, "py.test $cmdlineString") as CommandLineFile
+    }
+
+    /** Pytest plugins explicitly loaded/excluded by `-p plugin`/`-p no:plugin` in `addopts` */
+    open val loadedPlugins: List<PytestLoadPlugin> by lazy {
+        cmdlineFile.options
+            .filter { it.optionName == "-p" }
+            .mapNotNull { PsiTreeUtil.getNextSiblingOfType(it, CommandLineArgument::class.java) }
+            .map {
+                val name = it.valueNoQuotes
+                if (name.startsWith("no:")) {
+                    PytestLoadPlugin(name.substring(3), true)
+                } else {
+                    PytestLoadPlugin(name)
+                }
+            }
+    }
+
     companion object {
         const val CONFIG_PYTHON_CLASSES = "python_classes"
         const val CONFIG_PYTHON_FUNCTIONS = "python_functions"
+        const val CONFIG_ADDOPTS = "addopts"
 
         const val DEFAULT_PYTHON_CLASSES = "Test*"
         const val DEFAULT_PYTHON_FUNCTIONS = "test_*"
 
-        fun parse(file: VirtualFile?): PyTestConfig? {
+        fun parse(file: VirtualFile?, project: Project): PyTestConfig? {
             if (file == null) return null
 
             return when (file.extension) {
-                "ini" -> PyTestIni(file)
-                "toml" -> PyTestPyProjectToml(file)
+                "ini" -> PyTestIni(file, project)
+                "toml" -> PyTestPyProjectToml(file, project)
                 else -> null
             }
         }
@@ -252,11 +293,14 @@ internal fun Ini.getValueOrDefault(sectionName: String, optionName: String, defa
 /**
  * Parse a pytest.ini file and expose its contents
  */
-class PyTestIni(pytestIniFile: VirtualFile): PyTestConfig() {
+class PyTestIni(pytestIniFile: VirtualFile, project: Project): PyTestConfig(project) {
     private val pytestIni: Ini by lazy { Ini().read(BufferedReader(InputStreamReader(pytestIniFile.inputStream))) }
 
     override val pythonClassesRaw: String? by lazy { pytestIni.getValueOrDefault(PYTEST_INI_SECTION, CONFIG_PYTHON_CLASSES, null) }
     override val pythonFunctionsRaw: String? by lazy { pytestIni.getValueOrDefault(PYTEST_INI_SECTION, CONFIG_PYTHON_FUNCTIONS, null) }
+
+    override val cmdlineString: String by lazy { pytestIni.getValueOrDefault(PYTEST_INI_SECTION, CONFIG_ADDOPTS, null) ?: "" }
+    override val cmdlineOpts: List<String> by lazy { ParametersListUtil.parse(cmdlineString, false, true) }
 
     companion object {
         const val PYTEST_INI_SECTION = "pytest"
@@ -267,7 +311,7 @@ class PyTestIni(pytestIniFile: VirtualFile): PyTestConfig() {
 /**
  * Parse a pyproject.toml file and expose its tool.pytest.ini_options section
  */
-class PyTestPyProjectToml(pyprojectTomlFile: VirtualFile): PyTestConfig() {
+class PyTestPyProjectToml(pyprojectTomlFile: VirtualFile, project: Project): PyTestConfig(project) {
     private val pyprojectToml: TomlParseResult by lazy { Toml.parse(pyprojectTomlFile.inputStream) }
 
     override val pythonClasses: Regex by lazy { parseWildcardPatternsFromToml("$PYPROJECT_PYTEST_SECTION.$CONFIG_PYTHON_CLASSES", DEFAULT_PYTHON_CLASSES, true) }
@@ -277,7 +321,7 @@ class PyTestPyProjectToml(pyprojectTomlFile: VirtualFile): PyTestConfig() {
         return if (pyprojectToml.isArray(dottedKey)) {
             val patternsArray = pyprojectToml.getArray(dottedKey)!!
             if (patternsArray.containsStrings()) {
-                convertWildcardPatternsToRegex(patternsArray.toList() as List<String>, withDashes)
+                convertWildcardPatternsToRegex(patternsArray.toList().filterIsInstance<String>(), withDashes)
             } else {
                 // In the case of a non-string array (or heterogeneous array), pytest will fail.
                 // We instead simply treat the pattern as unmatchable.
@@ -288,6 +332,26 @@ class PyTestPyProjectToml(pyprojectTomlFile: VirtualFile): PyTestConfig() {
             convertWildcardPatternsStringToRegex(pyprojectToml.getString(dottedKey)!!, withDashes)
         } else {
             convertWildcardPatternsStringToRegex(defaultPatterns, withDashes)
+        }
+    }
+
+    override val cmdlineOpts: List<String> by lazy {
+        val dottedKey = "$PYPROJECT_PYTEST_SECTION.$CONFIG_ADDOPTS"
+        if (pyprojectToml.isString(dottedKey)) {
+            ParametersListUtil.parse(pyprojectToml.getString(dottedKey)!!, false, true)
+        } else if (pyprojectToml.isArray(dottedKey)) {
+            pyprojectToml.getArray(dottedKey)!!.toList().filterIsInstance<String>()
+        } else {
+            emptyList()
+        }
+    }
+
+    override val cmdlineString: String by lazy {
+        val dottedKey = "$PYPROJECT_PYTEST_SECTION.$CONFIG_ADDOPTS"
+        if (pyprojectToml.isString(dottedKey)) {
+            pyprojectToml.getString(dottedKey)!!
+        } else {
+            super.cmdlineString
         }
     }