+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project DOES NOT adhere to [Semantic Versioning](https://semver.org/spe
 [To Be Released]
 ----------------
 
+**Features:**
+ - Add completion/references for external lambda fixtures (from appropriate `conftest.py` files, and from plugins listed in the root conftest's `pytest_plugins` and from `-p <plugin>` arguments in the pytest config's `addopts`)
+
 _To Be Released..._